website.json: Initial website specs as Swagger 2.0 JSON
The Swagger specs are here [1], and I'm basing this on the
petstore-expanded.json sample [2].  We don't currently have terms or a
license of the API, so I'm leaving out the optional termsOfService and
license fields [3].  I've also used customer service as the contact
information, but we might want an API-specific contact.

For the host and basePath, I've gone with api.azurestandard.com.
Discussion from IRC on 2014-10-01:

  10:24 < shofetim> api.azurestandard.com or
    azurestandard.com/api.... Just FYI, doesn't look like you've
    gotten that far yet. The website will be one consumer of the API,
    3rd parties, and a mobile app will be others.
  ...
  10:28 < wking> +1 for api.azurestandard.com :)
  10:30 < shofetim> yes, my only concern is the pain we have
    experienced with CORS
  10:30 < wking> what pain?  I thought we had CORS straigtened out
    now?
  10:31 < shofetim> sure, but it seems to crop up often...
  10:31 < shofetim> *shrug* it is likely the better way to go.

We also mostly serve HTTPS now, but I don't have a problem with folks
using HTTP if they want to, as long as we use HTTPS for any endpoints
that require auth.

[1]: https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md
[2]: https://github.com/swagger-api/swagger-spec/blob/7d6fc05704f99aeddac3fced8822ce952cf87d1f/examples/v2.0/json/petstore-expanded.json
[3]: https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#fixed-fields-1