public.json: Rename from 'website.json'
"Website" is so vague ;).  Swagger does have some suggestions for
serving the API based on client permissions [1], but I think it's
better to keep the public API in a static file so users can track it
in version control and give us feedback.  I'd handle private paths or
path items by maintaining the internal API as a private branch of the
public API.

[1]: https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#security-filtering