public.json: Add drop.fees and a dropFees schema
The new website UI needs these to display while folks are looking for
drops to join or viewing information about a particular drop.
Internally, we keep these as references to strings, but to make the
API more flexible, we're boiling that down to two enumerated
attributes:

a. When do you have to pay a fee (drop.fees.condition)?
   1. Whenever you order on that drop (order).
   2. Whenever someone holds your order for you (held).
   3. Never (null).
b. How do you pay the fee (drop.fees.medium)?
   1. You don't have to pay any fees (null).
   2. With volunteer time (volunteer).
   3. With money (payment).

Here's how our internal strings will be rendered:

* No, this is a community drop.
  -> {"medium": null, "condition": null}
* No, but tips are welcome.
  -> {"medium": "payment", "condition": null}
* No, but volunteering to help is necessary.
  -> {"medium": "volunteer", "condition": "order"}
* Yes, for all orders.
  -> {"medium": "payment", "condition": "order"}
* Yes, but only if I have to hold orders.
  -> {"medium": "payment", "condition": "hold"}

but other combinations may occur in the future.  For example:

* {"medium": "volunteer", "condition": "hold"}
  -> "You have to volunteer if I hold an order for you."
* {"medium": "volunteer", "condition": null}
  -> "Volunteers welcome!"