public.json: Allow negative limits for tail-based slicing
The driving use-case is listing the next n trips that might visit a
drop.  Since trips are sorted in order of decreasing cutoff times,
we'll want to get the last n (oldest) trips that satisfy some cutoff
criteria (>= now).  We don't have a parameter for the cutoff criteria
yet, but the negative limits allow easy consumption from either side
of a query.

I've adjusted all the limit parameters.  The orderings for the other
queries are less clear, but I think it's best to maintain consistent
interpretations for a given parameter across endpoints.