public.json: Add 'PUT /drop/{id}' to update a drop
The data should match the usual 'drop' schema.  In this case, I don't
mind if folks set the 'id' property or not (whichever is easiest), but
if they set the 'id' property it must match the ID used in the URL.

Returning the updated drop doesn't seem useful (it should be
semantically equivalent to the uploaded data), so I've gone with '204
No Content' as the default success response.