public.json: Add 'PUT /route/{name}' to update a route
The data should match the usual 'route' schema.  In this case, I don't
mind if folks set the 'name' property or not (whichever is easiest),
but if they set the 'name' property it must match the name used in the
URL.

Returning the updated route doesn't seem useful (it should be
semantically equivalent to the uploaded data), so I've gone with '204
No Content' as the default success response.