public.json: Add 'GET /products?ancestor-category=...'
The categories are about filtering, not finding.  So in the cheese
category, we want to show all the cheeses.  In the cheese->cheddar
subcategory, we want to limit to only cheddars.  If the cheddars are
associated with cheese->cheddar (or cheese->cheddar->sharp), we still
want to list them in the cheese category (even though they aren't
directly associated with that category).