public.json: Add 'GET /addresses' and 'GET /address/{id}' endpoints
People can have multiple addresses associated with them, so we need
separate endpoints to manage those addresses.  Because other models
are also associated with addresses (e.g. drops), we're leaving the
person association off of the address model.

Because of the difficulty with adding a field for a few endpoints in
Swagger, I'm leaving the address model's 'id' field undocumented for
now.  For more details, see 03e139a0 (public.json: Add 'POST /drops'
to create a new drop, 2014-12-19).