public.json: Adjust products to key on ID, and not packaging code
The new products are generic, packaging-independent items (e.g. "Azure
Farm Unifine Barley Flour"), and the product-packaging instances are
specific sizes of that product (e.g. 5 lbs., 8 lbs., 19 lbs., 4x19
lbs., and 25 lbs.).

Previously we listed products by packaging-code, with
product.repackaged pointing at alternative packaging.  A good number
of links are per-packaging (categories, order-lines, ...), so we still
want a code-keyed interface.  However, we also want to be able to list
per-generic-product items for search results.  For example, you should
be able to search for "Unifine Barley Flour" and get a *single* item
with all possible sizes, instead of a separate item for each size.  It
would be possible for the client to use the old API and collate sizes
on their side, but that doesn't work well with pagination (you'd have
to fetch and collate *all* the entries, not just the ones on the
current page).  The new per-generic-product API lets you easily
paginate with per-generic-product items, and you can still use the old
per-product-packaging API to paginate with per-packaging items.