public.json: Use strings for credit-card numbers and security codes
The card number and security code can both be variable numbers of
digits:

* An ISO/IEC 7812 card number is most commonly 16 digits in length,
  and can be up to 19 digits [1].
* The securitry code can be three digits (MasterCard, Vida, Discover)
  or four (American Express) [2].

Both can be zero-padded:

* The first digit of a credit card number is the Major Industry
  Identifier (MII) (see ISO/IEC 7812), which represents the category
  of entity which issued the card. MII digits represent the following
  issuer categories: 0 - ISO/TC 68 and other future industry
  assignments [1].
* I've seen a physical card with a zero-padded security code.

We could pass these numbers around as numbers, and then pad with zeros
before submitting the information to our card processor, but without a
consistent length we don't know how many digits we should pad to (and
our processor needs the zero-padded version).  This commit works
around all of that by passing the "numbers" around as opaque strings.

[1]: http://en.wikipedia.org/wiki/Bank_card_number
[2]: http://en.wikipedia.org/wiki/Card_security_code