public.json: Add 'GET /route-stops' and /route-stop/{id} endpoints
We only create stops at verification time, but folks will want to have
estimated delivery times while they're shopping for their order.
David confirmed that a single model should store the usual offset
between the start of delivery and a given drop on a route, so we need
to expose that data publically.  This commit specifies that API and
adjusts the stop description to explain that it's only useful for
verified trips.

The duration format specifies ISO 8601 durations [1,2].  I've proposed
it for inclusion in future versions of Swagger [3], but until that's
accepted we're using it as a custom format [4].

[1]: https://tools.ietf.org/html/rfc3339#page-13
[2]: http://en.wikipedia.org/wiki/ISO_8601#Durations
[3]: https://github.com/swagger-api/swagger-spec/issues/359
[4]: https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#data-types
     Swagger uses several known formats to more finely define the data
     type being used. However, the format property is an open
     string-valued property, and can have any value to support
     documentation needs. Formats such as "email", "uuid", etc., can
     be used even though they are not defined by this specification.