public.json: Add address.name
For several mailing applications (e.g. mailing to P.O. boxes), we'll
need a contact name that may not match the associated customer's
person.name.  Even when the contact name is not required for delivery,
it can be useful to have an alternative contact name for shipping
orders to friends and relatives.  E.g.:

  {
    "name": "Dad",
    "street-address": "123 Example Rd.",
    ...
  }

This field gives you that flexibility.

I don't think it's useful for:

* Drops, where we should be using the main coordinator's name.
* Pickups, which are basically inverse drops.
* Credit cards, which already have a billing name.  We can relax this
  restriction if we ever hear of a credit card with one name for the
  card and another for the billing address.

On freenode's #microformats, tantek suggested using the more elaborate
http://microformats.org/wiki/h-card for name + address.  In our API,
the h-card role is filled by "person", so non-login person entries in
the address manager would be a possible alternative.  But the current
feeling (Tom, Casey, Jordan, and me) is that extending the address
model is simpler than supporting non-login people, so we're taking
this approach.