public.json: Convert trip.delivery-start from a date to a date-time
The date version landed in d2a35b36a (public.json: Add 'GET
/route-stops' and /route-stop/{id} endpoints, 2015-05-07), but
"stop-local time" is not something that makes sense for a trip
attribute (the first stop on a trip may not even be public).  Instead
just let the API implementation use a full date-time format to specify
their desired timezone explicitly.  This will leak a bit more
information about that first stop, since folks can use the target
delivery time to guess it's location or to guess which of several
closed drops is getting a delivery from a given trip, but those aren't
major leaks.

The main use of this field is to calculate target stop times using
route-stop.deliver-offset, so it doesn't really matter what event
we're using as a baseline.  We could avoid leaking the first stop's
target time by switching to a trip.ship time (when the truck leaves
the warehouse).  We can't do that now though, since we don't currently
track or schedule the departure times.