public.json: Add packagedProduct.favorite boolean and annotate-person parameter
Also rename the old 'person' parameter to 'filter-person' to avoid
confusion with the new parameter.

We want to be able to mark your favorites while you shop, without
requiring you to fetch all your favorites in a separate request.  We
also want to be able to show information like this when you're working
on behalf of someone else (e.g. you're a customer-service
representative shopping for a customer, or you're just trying to
reproduce an annotation-related bug that someone else is seeing).

The generic filter- and annotate- prefixes will let us consistently
extend this sort of handling to other fields besides people in the
future.

See the 2015-06-04 Slack discussion from around 2:36 to 4:10pm Pacific
for more details [1,2].

[1]: https://azurestandard.slack.com/archives/general/p1433453776001084
[2]: https://azurestandard.slack.com/archives/general/p1433459448001213