public.json: Add a telephone definition
RFC 2462 [1] and RFC 6350 [2] define a telephone entry model like
this (following RFC 6350 here).  The type enums mean:

text: supports SMS messages
voice: audio support
video: video conferencing support
fax: facsimile support
pager: paging device
textphone: for people with hearing or speech difficulties
cell: cellular or mobile
home: associated with a residence
work: associated with a place of work

The examples given in RFC 6350 for the author are [3]:

  {
    "pref": true,
    "type": ["work", "voice"],
    "number": "+1 418 656 9254 ext. 102"
  },
  {
    "type": ["work", "cell", "voice", "video", "text"],
    "number": "+1 418 262 6501"
  }

In both RFCs, the phone numbers are associated directly with the
person [4] and not part of the address [5].  But we may want a bit
more flexibility in our API so I'm adding this object in a separate
commit.

[1]: https://tools.ietf.org/html/rfc2426#section-3.3.1
[2]: https://tools.ietf.org/html/rfc6350#section-6.4.1
[3]: https://tools.ietf.org/html/rfc6350#section-8
[4]: https://tools.ietf.org/html/rfc6350#section-3.3
[5]: https://tools.ietf.org/html/rfc6350#section-6.3.1