public.json: Add registration endpoints
I've gone with POSTs for each endpoint and have avoided GETs, because
RFC 2616 says [1]:

  In particular, the convention has been established that the GET and
  HEAD methods SHOULD NOT have the significance of taking an action
  other than retrieval.  These methods ought to be considered "safe".
  This allows user agents to represent other methods, such as POST,
  PUT and DELETE, in a special way, so that the user is made aware of
  the fact that a possibly unsafe action is being requested.

And we don't want to use PUT, because the token is ephemeral and [2]:

  The PUT method requests that the enclosed entity be stored under the
  supplied Request-URI.

That leaves POST, which among several supported actions includes [3]:

  Providing a block of data, such as the result of submitting a form,
  to a data-handling process;

and:

  The action performed by the POST method might not result in a
  resource that can be identified by a URI. In this case, either 200
  (OK) or 204 (No Content) is the appropriate response status,
  depending on whether or not the response includes an entity that
  describes the result.

  If a resource has been created on the origin server, the response
  SHOULD be 201 (Created) and contain an entity which describes the
  status of the request and refers to the new resource, and a Location
  header (see section 14.30).

Which is why /registration/register returns a 200 with the resend
token and /registration/resend returns a 204.

Perhaps /registration/confirm should return a 201 with a Location
header pointing at /person/{id}, but all of our other POST-to-create
endpoints return a 200 and the created object directly, so that's what
we're doing here.

We need two tokens (one for resending and another for confirmation)
because the resend token gets sent to the client app, but we don't
want the client app to be able to confirm the user without the user
getting the token from their email (otherwise what's the point of
sending a confirmation email?).

We're going to want a company field at some point, but I've left that
off for now because we don't have anywhere to store it in the
database.

[1]: http://tools.ietf.org/html/rfc2616#section-9.1.1
[2]: http://tools.ietf.org/html/rfc2616#section-9.6
[3]: http://tools.ietf.org/html/rfc2616#section-9.5