public.json: Add updatePerson
We collect several person-related attributes when registering new
people (usually customers).  Instead of duplicating schema between
parameters in POST /registration/register and the coming PUT
/person/{id}, collect that new/updated-person schema in a definition
and use *that* in the pushed body.

For silly examples of this approach, see the current updateDrop and
updateRoute. For a bit more meat, see the spec around newPaymentMethod
and paymentMethod. For previous comments about why I'm cheating for
updateDrop and updateRoute, see 03e139a, 168a97b, 1616442, and
27ba2a1. In the route and drop cases, I've slightly broke the spec (by
including nominally requiring 'id').

I've also added company to person [1] and updatePerson and password to
updatePerson.

UPS customers who do not enter a postal code (which we use to estimate
shipping) will be prompted with a shop-time notification.  This is the
same handling we'll give non-UPS customers who decide to open a new
UPS cart.

The only hole in the loose requirements is if a user registers with
can-email unset, no phone numbers, and no mailing address.  In that
case, we have no way to push notifications to the customer.  But I'm
ok with that, because we'll get a billing address for most users when
they checkout their first order with a credit card.  For users that
wiggle into non-credit-card terms, we can make sure customer service
collects a contact channel while assigning the terms.  I'm ok with a
subset of customers that have never placed an order or negotiated for
non-credit-card terms being uncontactable.

[1]: https://docs.google.com/document/d/1dwayyitTQT5ma-UhFfkC16i6aVrJ2ArV-HnUdWyst10/edit#heading=h.acaorddfccxp
     10.2.3 Step #3: Join Azure
     ... This screen should allow the customer to enter the following
     information before clicking on join ...
     ...
     * Company Name (optional)
     ...