public.json: Add order.placed, .shipped, and .delivered times
We want to display these times when showing user's the current status
of their order.  For orders we deliver to drops, we can find the
delivery time by looking at the stop.time for the associated stop.
But for orders delivered by UPS or another carrier, we'll want to
cache delivery times that we get from the carrier here.