public.json: Use only one body parameter for login
The payload is simple enough that I've used an inline schema, rather
than pulling it out to a definitions.login object.

I've also fixed 'authorization' -> 'authentication', because these
fields are used to establish a user's identity (authenticating the
user), and then that identity is used (by the API implementation) to
authorize future requests.