public.json: Use only one body parameter for resendRegistrationEmail
This is a small object, but I've used a definition instead of an
inline schema to keep the base-url source close to
registration.properties.base-url, in the hopes that folks notice the
lack of DRYness and update both locations if they touch them in the
future ;).

I've also changed some "new person" language to use the more compact
"registrant" instead.