public.json: Add address.preference
We need a consistent way to select a user's primary address, and the
cleanest API for that seems to be letting the user order all their
preferences.  Then setting a new primary address would look something
like (using azure-angular-providers [1]):

  newPrimary.preference = 100;
  oldPrimary.preference = 0;
  newPrimary.$save().$promise.then(function () {
    oldPrimary.$save();
  });

There's a low probability that the newPrimary.$save goes through and
the oldPrimary.$save does not.  If the oldPrimary.preference was
originally 100, that means there is some time (while the
oldPrimary.$save is in flight, and possibly after if that call fails)
where both addresses will have the same preference on the server.  If
that's a concern, you could use something like:

  newPrimary.preference = oldPrimary.preference + 1;
  oldPrimary.preference = 0;
  newPrimary.$save().$promise.then(function () {
    oldPrimary.$save().$promise.then(function () {
      newPrimary.preference -= 1;
      newPrimary.$save();
    });
  });

If you weren't worried about rolling over the int32 (you're probably
safe ;), you could skip the decrement step.

[1]: https://github.com/azurestandard/azure-angular-providers