public.json: Require name, email, and password for registrants
Folks using updatePerson in other contexts are free to only set the
fields they intend to update, but for creating new people we need an
email address (to send the confirmation email) and a password (so they
can log in after confirming).  And we might as well have a name ;).