public.json: List the default values for boolean person creation
The 'registerPerson' endpoint uses the 'registration' model, which
includes the 'updatePerson' model.  To avoid folks having to guess
what happens when they don't specify a boolean field, this commit
explicitly lists the default value for those fields.

If a field is not specified during an 'updatePerson' call, we leave it
alone.