public.json: Add person.price-level and a multi-level price object
This is a first step in our shift to more flexible pricing.  The
current API implementation checks the requester's price level and
returns the appropriate prices object (now _price, after this
commit). As we shift to Algolia for product searches, we'll need less
dynamic pricing information, because Algolia isn't going to
authenticate reqeusters for us.  This commit associates
requester-independent pricing information with the packaged-products,
and gives customers the information they need (via their /person or
/person/{id} result) to select the appropriate price from the satic
packaged-products information.

In the future, we will likely transition to a more
complicated/flexible pricing model, with base-price information on
packaged-products that's not price-level dependent (e.g. a base price
and scaling coefficient).  Then the per-person information will
include a formula for converting that person-independent information
into a price.  But this is too fuzzy to move toward at the moment, and
we need the person-independent information now ;).