public.json: Shift POST /payments?payment-method=...&amount=... to body
We're going to be using this model for an order attribute as well, so
moving it to the definitions section will allow us to avoid
duplication.

The parameters were initially query parameters to get easy logging
(see 2b8dbfeb, public.json: Add POST /account-entries, 2015-08-28),
but now that we have a plan for logging payload JSON that's not
important anymore.