public.json: Add order.checkout-payment
So customers can associate a payment method (e.g. a credit card) to be
used for paying off the order (by PUTing or POSTing orders with
checkout-payment.payment-method set).  After shipping,
checkout-payment will just become a record of whether or not they
associated a payment method with the order (e.g. they may have not
associated a payment method, and just covered the order with previous
credits from POST /payments or some such).