public.json: Add order.fees
Existing fees in the database look like:

  {"order": 123, "type": "shipping", "amount": 0.39, "notes": "8.50%"}
  {"order": 456, "type": "small-order", "amount": 2.50,
   "notes": "Automatically Added"}

I'm not wild about the "Automatically Added" (do we have fees that are
manually assessed?), but having room for an optional note of some kind
makes sense.

Shipping and small-order fees are the only ones I know of, but we can
easily extend this to new fees by adding entries to orderFeeType.

The explicit orderFee.id, orderFee.order, and use of an array for
order.fees instead of an object are setting us up for an /order-fee
endpoint if we decide to move fee management into the API
(e.g. letting internal UIs POST fees to orders, or DELETE fees from
orders).

I haven't required orderFee.id at the moment, because we currently
handle UPS shipping outside of our fee model.  If we do decide to add
an API for manipulating order-fee instances we'll need to collapse
that into our existing order-fee table.