public.json: Add 'GET /purchases' and 'GET /purchase/{id}'
Trucks on routes visit stops to drop off orders and/or pickup
purchases.  Logistics needs to organize pickup locations along routes
to plan trips, and these endpoints (based on the similar order
endpoints) are one link in that chain.  This commit also introduces
references to, but does not yet define, the following models:

* vendor: a company that sells products to Azure.
* pickup: a location used for repeat purchase pickups, just
  as a drop is a location used for repeat purchase delivery.

We also receive purchases via common carrier or other arangements, but
I haven't modeled that yet because logistics only cares about backhaul
pickups that they need to schedule.

The status enum is a bit more detailed than order.status because
there's more back-and-forth between Azure's buyers and vendors than
there is between customers and Azure.  The lifecycle is:

* open: the purchase order is being formed internally
* submitted: the purchase order has been sent to the vendor
* confirmed: the vendor and Azure agree on quantities, pricing, and
  logistics.
* shipped: the purchased stock has left the vendor's location.
* delivered: the purchased stock has arrived at Azure's location.
* labeled: the purchased stock has been labeled and is ready to be put
  away.  Damaged stock has been identified for return to the vendor.
* reconciled: any damaged stock has been returned to the vendor,
  missing stock has been accounted for with the vendor, and agreement
  has been reached on the final price.
* paid: Azure has

Compared with order.status, there are no 'open', 'placed', 'cart', or
'lost', or 'labeled' analogs; 'confirmed' is analagous to 'placed';
and 'shipped' and 'delivered' are the same.

Where we store a status-change date-time in the database, I've also
added a field to reflect that time in the API model.  Besides giving
some information about the purchase's history, this sort of approach
also simplifies status modeling.  You can see if a stage has been
completed by checking for its time, instead of remembering an internal
ordered list of statuses in each client.  Unfortunately, we don't
currently track all the times we'd need to drop the status field
entirely.