public.json: Add 'GET /pickups' and 'GET /pickup/{id}'
Based on /drops and /drop/{id}, although I've only copied the read
endpoints for now.

I'm not sure if we have any pickup locations now that are supplied by
multiple vendors, but I can imagine that happening at a multi-modal
hub or other location that served as a middleman.  Regardless of
whether or not we do that *now*, it seems like a useful situation to
allow in the API, and it doesn't make things much more complicated
(you just need a separate membership endpoint to associate vendors
with pickups, just like we need a separate endpoint to associate
customers with drops).