public.json: Add 'GET /vendors' and 'GET /vendor/{id}'
Testing the pickup and purchase APIs is easier if we can go from
manager -> vendor -> {pickups and purchases}.  This commit stubs in
enough of our internal vendor representation to accomplish that, given
a manager's ID:

  for vendor in /vendors?manager={manager-ID}:
    /pickups?vendor={vendor.id}
    /purchases?vendor={vendor.id}

To add a bit more context around manager vs. buyer, managers are users
who can edit the vendor's URL, etc. on behalf of the vendor (and they
are usually vendor employees, but don't have to be).  Buyers are Azure
employees who each manage a collection of vendors, so vendors can have
a consistent point of contact within Azure to avoid hopping around
within our buyer pool.