public.json: Add 'POST /password/reset' and 'POST /password/confirm'
Give folks a way to reset their password.  POSTing a known email
address to /password/reset will cause a message to be sent to that
address.  The email will contain a link with a reset token, and that
token can be POSTed to /password/confirm to set a new password.  This
base-url / token approach is just like the one used for registration,
and we support both POST and GET confirmations for the same reasons
(see 182d6c9, public.json: Add 'GET /registration/confirm?token=...',
2015-09-03, #19).

/password/confirm returns the person whose password was reset, so the
UI can display their name, etc. in a successful-reset page.