add product data as an optional attribute to packaged-product
The name of the product and the name of the brand are required for each
packaged-product in the internal-website categories app. Including them
in the packaged-product response prevents the need to send another api
call to /products for each packaged product. When loading a large
number of packaged-products (such as the uncategorized
packaged-products), the browser can be overwhelmed with the number of
extra api calls. This prevents that.