add "primary" field to the category model
Requied in order to accommodate the primary category feature. (see
azurestandard/website#86)

The requirements for the categories app include:
1. In the category app on the left hand side allow the user to mark "Categories" as primary.
2. When dragging a Product into a Category look and see if the Category is marked as primary.
a. If it is then mark the Category as primary for this Product.
b. If this Product already has a primary Category then display a dialog prompting the user to select which Category is primary.
c. If it is not then do not mark the Category as primary for this Product.

So we need to be able to mark a category as a "primary category"