public.json: Add person.permissions
Add a way to retrieve permissions for your authenticated user
(/person) or someone else (/person/{id}).  This currently has very
coarse permissions for whether the person is staff or a superuser
(with all permissions), so we can do things like only exposing the
internal-website UI to staff [1].  Moving forward, we can fill this in
with entries like 'create-address', 'update-address', and
'delete-address' for folks who can CrUD /addresses, etc. [2].  That
level of granularity will take time, but this setting will let us add
per-endpoint permissions and migrate per-UI checks as we see fit.

There's some discussion in [1] about where to put this information,
with /session and a new /permissions being floated.  I've gone with
/person here, because:

* It's an endpoint we were hitting already, so clients don't need a
  separate request.  Filling this in with permission data shouldn't
  make the endpoint much slower.

* It allows authorized users to request permission for other people,
  which can help answer questions like "why can't Alice view the
  logistics UI".

[1]: https://github.com/azurestandard/internal-website/issues/48
[2]: https://github.com/azurestandard/internal-website/issues/48#issuecomment-179975949