public.json: Add ?persist=true to stop POST/PUT/DELETE
You could handle this push in the client with a few calls to stop
endpoints and a call to a route-stop endpoint.  But after talking this
over with Tom and Jeff yesterday, the consensus was that we should
bake David's preferred the business logic [1] into the API.  I've gone
with a query parameter since that seemed like a smaller API footprint
than a parallel set of endpoints.  And clients that do want to handle
the persistence logic on their own can just not set the query
parameter.

As a concrete example of the persistence the query parameter enables,
say route A1 has trips 12 (cutoff March) and 34 (cutoff April).
Adding a stop for drop 56 on trip 12 would also add a stop for that
drop with the same offset to trip 34 and to the A1 route.  Adding drop
78 to trip 34 would affect trip 34 and add the drop to the A1 route,
but would not put drop 78 on trip 12.

If you want to add a stop to a route but not to any existing trip, you
still need to use the route-stop interface directly.

[1]: https://github.com/azurestandard/internal-website/issues/51#issuecomment-181720704